Blow it all away
I started this project on May 13, 2012 and published it in a month. I
had no idea how to write Ruby or Python code, and I was new to a lot of
the concepts that went along with building a website. It has been over
three years since then, and I am a lot more experienced in such matters.
It helps to be gainfully employed as a Rails developer.

I am proud of what I accomplished in building this site, but I am no
longer proud of the code that composes it. It is frustrating to
maintain, and I dread any time I have to plunge back into the code base.

With that, I am starting over. This time it's a Rails app, and I plan on
taking a better approach to things (like, you know, having tests). Time
to clear out the cruft.