Change generator template to include example tests
It was getting tedious writing almost identical tests for every
generated link object. Modified the test template to include assertions.
This gets most of the way to an actual test; just fill in the expected
value.

Add skip statements by default so that we don't get failed assertions
before finishing the test.