Add 'selected' css class logic to presenters
The view needs to distinguish which season, team, and game is currently
displayed. This is done by adding a 'selected' css class to the
appropriate <li> tags.

Added #li_classes methods to the various presenters to handle the logic
of determining which <li> is selected. The views will loop through this
array the same as it does the array returned from the #links methods.