Ignore tests in Links module when generating links
The ExternalLinkPresenter loops through all classes defined in the Links
module and instantiates them. Unit tests are included in this module as
well.

The `#sort` method blew up in ExternalLinkPresenter because it was
comparing `Links::Base` objects with `ActiveSupport::TestCase` objects.

This was only a problem when running the entire test suite, not when
actually using the site.