Add another regular season game fixture
This is in preparation for testing the game ordering feature. In order
to test how games are sorted, we need more than one game fixture.

This broke some tests, which are fixed in this commit as well. I'm
starting to see the brittleness of fixtures. I might switch to factories
if this happens again.