Add miscellaneous links
Adding a third section of links to the site. This section is for sites
that don't break up their content by season and/or game. This allows me
to link to sites that don't allow passed in parameters or simply don't
filter data with parameters.

There are more misc. links to come. This is just the first pass.