Add rake task that checks for missing games
This task calls out to a GameInspector object. This object loops through
all games in the database for a particular season and compares it
against an ideal list of games. Any mismatches are reported as missing
games.

For now this only includes a check for regular season games. The
numbering scheme for playoff games is complicated.