Add interactive routes to api documentation page
Adds back a feature of NTOI v1's api page. Each api route is displayed
and has documentation. It can be clicked to expand a form that lets the
user enter parameters. When the GET button is clicked, the actual api
call is performed, and the result is loaded into a textarea.