Fix page setup when using turbolinks
The setup of JavaScript events was taking place inside a jQuery ready
event. This event isn't fired if the page is loaded with Turbolinks, for
example, if the user clicks the API link from the main page. The result
was an unusable page.

Moved the setup into its own function and call it from both the read
event and Turbolinks' `page:load` event.