Update game date if game already exists
Sometimes a game gets moved. Other times the game is shown on the
schedule page before the actual game date (this happened during the
Christmas break).

When this happens, the game scraper would fail to save the game because
it already exists. Now the behavior is to find the game that exists and
update its date.