Refactor GameSaver using Chain of Responsibility
After changing GameSaver to update game dates, the branching logic was
starting to look ugly.

Refactored to use the Chain of Responsibility pattern. Each step
attempts to save the game. If it is successful, it returns its result
message. If it is not successful, it calls the next step in the chain.

Hopefully this makes it easier to add to the game saving logic in the
future. If not, at least it was an interesting experiment in applying
this pattern.