Fix scraper for new NHL.com schedule page
On Feb. 4, 2016, the NHL redesigned their game schedule page. This broke
my scraper since the layout and class names are completely different.

The page now includes all games from the given date forward. If I want
only one day's worth of games, I have to only parse the first table in
the list.

Team abbreviations no longer appear on this page. To identify the
participating teams, I have to use the full name instead. This required
an update the GameBuilder.

There is also no longer a specific message displayed when no games are
scheduled for a given day. I have to check if the first date on the page
matches the current date.