Add cache_key method to RequestContext
Rendering certain parts of the main page is starting to take a while,
specifically the games partial. I keep getting alerts from New Relic. It
is time to use some view caching.

Need a cache key that encompasses all variables of the request so that
it is unique to the request. Also included the date since the content of
the page is updated daily.

I put this in RequestContext since it felt like the best home. It also
feels like an SRP violation, but I'm living with it for now.

Also allowed the date to be injected for testing purposes. I'm resisting
the urge to use something like Timecop.