Use null object pattern for game
There are scenarios where a game cannot be found for a given set of
parameters; for example, we're looking for playoff games for a team that
did not make the playoffs.

These scenarios have the potential for nil errors. Rather than put a
bunch of nil checks everywhere, I'm using the null object pattern. This
is partly a black hole null object. It only responds to a limited set of
game-related methods, but those methods return a black hole object that
allows arbitrary method calls.