Make NullGame behave 'falsey'
I am checking for the presence of a game in ExternalLinkPresenter.  The
null object pattern is supposed to remove the need for such checks, but
in this case I really want to bail out early from the method if there is
no game.

I decided to make NullGame respond to methods like `#present?` for this
purpose. I am still unsure of this decision since it seems not to align
with the spirit of the null object pattern. I may remove this later.