Add RequestContext#to_hash method
I want to store request contexts in the database using the new Stat
model. Adding to_hash makes it serializable in the same manner as the
params hash.

It is starting to feel like this class is doing too much. If it grows
much larger, consider refactoring. One refactoring idea is to pull out
the queries and make RequestContext a value object.