Add class that generates cache keys
Rendering certain parts of the main page is starting to take a while,
specifically the games partial. I keep getting alerts from New Relic. It
is time to do some view caching.

I need a cache key that encompasses all variables of the request so
it is unique to that request. Also included the date since the content of
the page is updated daily.