Add CacheKey object (again)
In my previous attempt at caching, I had different cache keys for each
fragment of the main page. I attempted to identify which pieces of the
request context were needed by each fragment and embed those pieces in
the cache key.

This turned out to be a mistake because each fragment of the page actually
needs the entire request context to uniquely identify it. That is
because each anchor on the page uses the entire request context in its
URL params (which needed to happen when I added Turbolinks).

This time around, I am added one cache key for the entire page that uses
the entire request context. It also includes today's date because the
page changes each day during the NHL season.