Change schedule URL for scraping
The NHL.com schedule page now loads the body of the table
asynchronously, which breaks the game scraper. The contents have not
changed once it's loaded, it's just not loaded right away.

Changed the URL to point to the new location of the schedule body. I
found this new URL using the network tab in browser dev tools.