Set UTF-8 encoding when parsing NHL.com
The special character in "Montréal" broke the scraper again. Nokogiri
uses the declared encoding of the document it parses. In this case, it
was no longer encoding as UTF-8, possibly because I changed the url it
uses in af2dfed. The é character was being parsed as &Atilde;&copy;
which could not be matched to a team in the database. Explicitly
declaring UTF-8 encoding fixes the issue.