Remove link models
This was an experiment, and it proved more trouble than it was worth. The
idea was to treat third-party site links like individual objects. It
ended up being too tedious to add and edit links. All these links are
now represented in new presenter objects, so we can remove them.