Refactor creation of link presenters
I never liked the idea of assigning every link presenter to an instance
variable. I only did it for expediency. Now that RuboCop is complaining
about it, it's time to refactor.

I moved the instantiation to another object called LinkPresenter whose
job is to provide instance of link presenters for each site. I don't
love the name of this class and may change it. There's also no unit
tests because I think this kind of thing is better handled by
integration tests.